attempt to type check
