+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/Annotator.java
@@ -236,6 +236,6 @@ public class Annotator {
    * </ul>
    */
   private void suppressRemainingErrors() {
-    new NullAway().suppressRemainingAnnotations(config, injector);
+    new NullAway(config).suppressRemainingAnnotations(config, injector);
   }
 }
