+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/checkers/CheckerBaseClass.java
@@ -25,7 +25,6 @@
 package edu.ucr.cs.riple.core.checkers;
 
 import com.google.common.collect.ImmutableSet;
-import edu.ucr.cs.riple.core.Checker;
 import edu.ucr.cs.riple.core.Config;
 import edu.ucr.cs.riple.core.metadata.Context;
 import edu.ucr.cs.riple.core.metadata.index.Error;
@@ -35,18 +34,15 @@ import edu.ucr.cs.riple.injector.location.OnField;
 import java.util.Set;
 
 /** Base class for all checker deserializers. */
-public abstract class DeserializerBaseClass implements CheckerDeserializer {
+public abstract class CheckerBaseClass<T extends Error> implements Checker<T> {
 
   /** Annotator config. */
   protected final Config config;
-  /** The checker that this deserializer is supporting. */
-  protected final Checker checker;
   /** The supporting serialization version of this deserializer. */
   protected final int version;
 
-  public DeserializerBaseClass(Config config, Checker checker, int version) {
+  public CheckerBaseClass(Config config, int version) {
     this.config = config;
-    this.checker = checker;
     this.version = version;
   }
 
@@ -61,7 +57,7 @@ public abstract class DeserializerBaseClass implements CheckerDeserializer {
    * @param context Context of the modules which errors are reported on.
    * @return The corresponding error.
    */
-  protected Error createError(
+  public T createError(
       String errorType,
       String errorMessage,
       Region region,
@@ -72,7 +68,7 @@ public abstract class DeserializerBaseClass implements CheckerDeserializer {
         resolvingFixes.stream()
             .filter(f -> !context.getNonnullStore().hasExplicitNonnullAnnotation(f.toLocation()))
             .collect(ImmutableSet.toImmutableSet());
-    return new Error(errorType, errorMessage, region, offset, fixes);
+    return createErrorFactory(errorType, errorMessage, region, offset, fixes);
   }
 
   /**
@@ -82,7 +78,7 @@ public abstract class DeserializerBaseClass implements CheckerDeserializer {
    * @param onField Location of the field.
    * @return The updated given location.
    */
-  protected OnField extendVariableList(OnField onField, Context context) {
+  protected static OnField extendVariableList(OnField onField, Context context) {
     Set<String> variables =
         context
             .getFieldRegistry()
@@ -92,12 +88,7 @@ public abstract class DeserializerBaseClass implements CheckerDeserializer {
   }
 
   @Override
-  public int getVersionNumber() {
+  public int getVersion() {
     return version;
-  }
-
-  @Override
-  public Checker getAssociatedChecker() {
-    return checker;
   }
 }
